pyproject.toml
[tool.poetry]
name = "wf-classroom-video-uploader"
version = "1.0.1"
description = ""
authors = ["Paul J DeCoursey <paul@decoursey.net>"]
license = "MIT"
readme = "README.md"
packages = [{include = "uploader"}]

[tool.poetry.dependencies]
python = "^3.8"
redis = ">=3.2.1"
PyYAML = ">=4.2b1"
minio = ">=4.0.17"
click = ">=7.0"
pytelegraf = ">=0.3.3"
wf-video-io = ">=3.2.1"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
